fix: security fixes and magic number extraction
Security:
- Fix integer overflow in visual capture buffer indexing (u16 * u16)
- Fix NaN panic in keyframe sorting (unwrap → unwrap_or)
- Fix unchecked f32→u16 cast in grid layout (add .min() clamp)

Magic numbers:
- Add XTERM_CUBE_DIM/BASE/STEP, XTERM_GRAY_BASE/STEP/COUNT,
  WEBSAFE_STEP constants to core/constants.rs
- Replace hardcoded 36/6/55/40/8/10/51 in ansi.rs and color_picker

+++ src/testing/visual/capture.rs
@@ -46,7 +46,7 @@ impl VisualCapture {
     /// Get cell at position
     pub fn get(&self, x: u16, y: u16) -> Option<&CapturedCell> {
         if x < self.width && y < self.height {
-            let idx = (y * self.width + x) as usize;
+            let idx = y as usize * self.width as usize + x as usize;
             self.cells.get(idx)
         } else {
             None
